Merges

+++ 1.5.06 with T-6638/xrGame/WeaponMagazined.h
@@ -80,20 +80,20 @@ public:
 	virtual void	net_Export		(NET_Packet& P);
 	virtual void	net_Import		(NET_Packet& P);
 
-	virtual void	OnH_A_Chield		();
+	virtual void	OnH_A_Chield	();
 
-	virtual bool	Attach(PIItem pIItem, bool b_send_event);
-	virtual bool	Detach(const char* item_section_name, bool b_spawn_item);
-	virtual bool	CanAttach(PIItem pIItem);
-	virtual bool	CanDetach(const char* item_section_name);
+	virtual bool	Attach			(PIItem pIItem, bool b_send_event);
+	virtual bool	Detach			(const char* item_section_name, bool b_spawn_item);
+	virtual bool	CanAttach		(PIItem pIItem);
+	virtual bool	CanDetach		(const char* item_section_name);
 
-	virtual void	InitAddons();
+	virtual void	InitAddons		();
 
-	virtual bool	Action			(s32 cmd, u32 flags);
+	virtual bool	Action			(u16 cmd, u32 flags);
 	bool			IsAmmoAvailable	();
 	virtual void	UnloadMagazine	(bool spawn_ammo = true);
 
-	virtual void	GetBriefInfo				(xr_string& str_name, xr_string& icon_sect_name, xr_string& str_count, string16& fire_mode);
+	virtual void	GetBriefInfo	(xr_string& str_name, xr_string& icon_sect_name, xr_string& str_count, string16& fire_mode);
 
 public:
 	virtual bool	SwitchMode				();
@@ -108,7 +108,7 @@ protected:
 	int				m_iQueueSize;
 	//количество реально выстреляных патронов
 	int				m_iShotNum;
-	//после какого патрона, при непрерывной стрельбе, начинается отдача (сделано из-зи Абакана)
+	//после какого патрона, при непрерывной стрельбе, начинается отдача (сделано из-за Абакана)
 	int				m_iShootEffectorStart;
 	Fvector			m_vStartPos, m_vStartDir;
 	//флаг того, что мы остановились после того как выстреляли
