+++ 1.5.06 with T-6638/xrGame/WeaponMagazinedWGrenade.h
@@ -46,7 +46,7 @@ public:
 	virtual void	OnEvent			(NET_Packet& P, u16 type);
 	virtual void	ReloadMagazine	();
 
-	virtual bool	Action			(s32 cmd, u32 flags);
+	virtual bool	Action			(u16 cmd, u32 flags);
 
 	virtual void	UpdateSounds	();
 
@@ -78,14 +78,15 @@ public:
 	CWeaponAmmo*			m_pAmmo2;
 	shared_str				m_ammoSect2;
 	xr_vector<shared_str>	m_ammoTypes2;
-	u32						m_ammoType2;
+	u8						m_ammoType2;
 	shared_str				m_ammoName2;
 	int						iMagazineSize2;
 	xr_vector<CCartridge>	m_magazine2;
+
 	bool					m_bGrenadeMode;
 
 	CCartridge				m_DefaultCartridge2;
-	int						iAmmoElapsed2;
+	u8						iAmmoElapsed2;
 
 	virtual void UpdateGrenadeVisibility(bool visibility);
 };
